show polling interval
